Set git username to "Name Surname" instead of alias
The GitHub web UI defaults to using "Name Surname" when applying
commits, therefore sometimes there could be mismatches when working on a
project if some commits are done directly from the web UI (updates to
the README for example). This is really annoying, and maybe sticking to
"Name Surname" is the best option anyway.